+++ lychee-bin/src/formatters/mod.rs
@@ -1,5 +1,6 @@
 pub(crate) mod color;
 pub(crate) mod duration;
+pub(crate) mod hints;
 pub(crate) mod host_stats;
 pub(crate) mod log;
 pub(crate) mod response;
